feature/ archway module setup (#57)
* feat: archway app_module and codec

* feat: archway setup

* feat: scaffold methods for archway provider

* feat: scaffold archway_chain_processor

* feat: archway scaffold finalize

* feat: archway provider and processor added to strategies

* chore: fix failed test

+++ relayer/chains/icon/codec_test.go
@@ -56,6 +56,6 @@ func TestCodecEncode(t *testing.T) {
 	if err != nil {
 		assert.Fail(t, "Couldn't unmarshal interface ")
 	}
-	assert.Equal(t, ptr.GetLatestHeight().GetRevisionHeight(), 40)
+	assert.Equal(t, ptr.GetLatestHeight().GetRevisionHeight(), uint64(40))
 
 }
